# specs/api/openapi.yaml
openapi: 3.1.0
info:
  title: TableFlow API
  version: 0.1.0
  description: >-
    Machine-readable wire contract for implemented TableFlow routes only.
    Policies and planned routes are described in http.md. Routes are added
    here before their implementation lands.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:8080
    description: Local Go API (also reachable through each frontend's same-origin /api/v1 proxy)
security: []
...
paths:
...
  /api/v1/sessions/staff:
    post:
      operationId: createStaffSession
      tags: [staff]
      summary: Staff sign-in
      description: >-
        Requires an allowed admin Origin. Throttled per email (10) and client
        IP (30) per 10 minutes. Unknown, inactive and wrong-password accounts
        receive the same 401. Sets the host-only `__Host-tf_staff` cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email: {type: string, maxLength: 254}
                password: {type: string, maxLength: 512}
      responses:
        '201':
          description: Session created
          headers:
            Set-Cookie:
              description: '`__Host-tf_staff=<token>; Path=/; Expires=...; HttpOnly; Secure; SameSite=Strict`'
              schema: {type: string}
            Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StaffIdentity'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '415': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /api/v1/sessions/current:
    get:
      operationId: getCurrentStaffSession
      tags: [staff]
      summary: Current staff identity and roles
      security: [{staffSession: []}]
      responses:
        '200':
          description: Authenticated staff member
          headers:
            Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StaffIdentity'}
        '401': {$ref: '#/components/responses/Error'}
    delete:
      operationId: deleteCurrentStaffSession
      tags: [staff]
      summary: Sign out (revoke this session)
      security: [{staffSession: []}]
      responses:
        '204':
          description: Session revoked and cookie cleared
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
  /api/v1/staff/activate:
    post:
      operationId: activateStaffAccount
      tags: [staff]
      summary: Set the password of an invited account with its single-use token
      description: >-
        Requires an allowed admin Origin; throttled per client IP (20 per 10
        minutes). Invalid, used, revoked and expired tokens all return 422
        TOKEN_INVALID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [token, password]
              properties:
                token: {type: string}
                password: {type: string, minLength: 12, maxLength: 512}
                display_name: {type: string, maxLength: 100}
      responses:
        '200':
          description: Account activated
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [status]
                properties:
                  status: {type: string, const: active}
        '400': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '415': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /api/v1/branches/{branch_id}/staff:
    parameters:
      - {name: branch_id, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      operationId: listStaff
      tags: [staff]
      summary: Page through a branch's staff accounts (manager)
      description: Other branches and malformed IDs return 404; non-managers get 403.
      security: [{staffSession: []}]
      parameters:
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 25}}
        - {name: cursor, in: query, schema: {type: string, maxLength: 256}}
      responses:
        '200':
          description: One keyset page ordered by (created_at, id)
          headers:
            Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StaffPage'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
    post:
      operationId: inviteStaff
      tags: [staff]
      summary: Create an invited staff account (manager)
      description: The activation token is returned once; share `/activate#<token>` privately.
      security: [{staffSession: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, display_name, roles]
              properties:
                email: {type: string, maxLength: 254}
                display_name: {type: string, minLength: 1, maxLength: 100}
                roles: {$ref: '#/components/schemas/Roles'}
      responses:
        '201':
          description: Invited account and one-time activation token
          headers:
            Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [staff, activation]
                properties:
                  staff: {$ref: '#/components/schemas/Staff'}
                  activation: {$ref: '#/components/schemas/Activation'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '413': {$ref: '#/components/responses/Error'}
        '415': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
  /api/v1/staff/{staff_id}/activation:
    parameters:
      - {name: staff_id, in: path, required: true, schema: {type: string, format: uuid}}
    post:
      operationId: reissueStaffActivation
      tags: [staff]
      summary: Replace an invited account's activation token (manager)
      security: [{staffSession: []}]
      responses:
        '201':
          description: New one-time token; the previous one is revoked
          headers:
            Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Activation'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /api/v1/staff/{staff_id}/roles:
    parameters:
      - {name: staff_id, in: path, required: true, schema: {type: string, format: uuid}}
    patch:
      operationId: setStaffRoles
      tags: [staff]
      summary: Replace roles (manager); revokes the target's sessions
      security: [{staffSession: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [expected_version, roles]
              properties:
                expected_version: {type: integer, minimum: 1}
                roles: {$ref: '#/components/schemas/Roles'}
      responses:
        '200':
          description: Updated account
          headers:
            Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Staff'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '415': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
  /api/v1/staff/{staff_id}/deactivate:
    parameters:
      - {name: staff_id, in: path, required: true, schema: {type: string, format: uuid}}
    post:
      operationId: deactivateStaff
      tags: [staff]
      summary: Disable an account (manager); revokes sessions and open tokens
      security: [{staffSession: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [expected_version, reason]
              properties:
                expected_version: {type: integer, minimum: 1}
                reason: {type: string, minLength: 1, maxLength: 500}
      responses:
        '200':
          description: Disabled account
          headers:
            Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Staff'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '415': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    guestSession:
      type: apiKey
      in: cookie
      name: __Host-tf_guest
      description: Capability-derived guest session on the PWA origin; valid only for the capability's current generation.
    staffSession:
      type: apiKey
      in: cookie
      name: __Host-tf_staff
      description: Host-only staff session cookie. Mutations also require an allowed admin Origin.
  responses:
    Error:
      description: Error in the common envelope
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    RateLimited:
      description: Too many attempts
      headers:
        Retry-After:
          description: Seconds until the window resets
          schema: {type: integer, minimum: 1}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  headers:
...
    PrivateNoStore:
      description: Authenticated responses are private and never stored.
      schema:
        type: string
        const: private, no-store
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id, fields]
          properties:
            code:
              type: string
              pattern: '^[A-Z][A-Z0-9_]*$'
              examples: [DEPENDENCY_UNAVAILABLE]
            message:
              type: string
            request_id:
              type: string
            fields:
              type: object
              additionalProperties:
                type: string
    Roles:
      type: array
      minItems: 1
      maxItems: 4
      uniqueItems: true
      items:
        type: string
        enum: [cashier, host, kitchen, manager]
    StaffIdentity:
      type: object
      additionalProperties: false
      required: [staff_id, branch_id, email, display_name, roles]
      properties:
        staff_id: {type: string, format: uuid}
        branch_id: {type: string, format: uuid}
        email: {type: string}
        display_name: {type: string}
        roles: {$ref: '#/components/schemas/Roles'}
        session_expires_at: {type: string, format: date-time}
    Staff:
      type: object
      additionalProperties: false
      required: [id, branch_id, email, display_name, status, roles, version, created_at]
      properties:
        id: {type: string, format: uuid}
        branch_id: {type: string, format: uuid}
        email: {type: string}
        display_name: {type: string}
        status: {type: string, enum: [invited, active, disabled]}
        roles: {$ref: '#/components/schemas/Roles'}
        version: {type: integer, minimum: 1}
        created_at: {type: string, format: date-time}
    StaffPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor, server_time]
      properties:
        items:
          type: array
          maxItems: 100
          items: {$ref: '#/components/schemas/Staff'}
        next_cursor: {type: [string, 'null']}
        server_time: {type: string, format: date-time}
    Activation:
      type: object
      additionalProperties: false
      required: [token, expires_at]
      properties:
        token: {type: string, description: 'Single-use secret; shown once'}
        expires_at: {type: string, format: date-time}
...